backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.9
bcrypt==5.0.0
boto3==1.42.73
botocore==1.42.73
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
fastapi==0.135.1
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jmespath==1.1.0
openpyxl==3.1.5
proto-plus==1.27.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
PyMuPDF==1.27.2.2
python-docx==1.2.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
